+++ src/aux-server/aux-web/aux-player/MenuBot/MenuBot.ts
@@ -74,6 +74,7 @@ import { formatModalityButtonId, Input } from '../../shared/scene/Input';
 import { SvgIcon } from '@casual-simulation/aux-components';
 import { Subscription } from 'rxjs';
 import type { BotManager } from '@casual-simulation/aux-vm-browser';
+import Bowser from 'bowser';
 
 @Component({
     components: {
@@ -108,6 +109,7 @@ export default class MenuBot extends Vue {
     hoverStyle: MenuBotResolvedHoverStyle = 'hover';
     cursor: string = null;
     alwaysShowSubmit: boolean = false;
+    inputMultiline: 'default' | boolean = 'default';
 
     private _down: boolean = false;
     private _hover: boolean = false;
@@ -127,7 +129,8 @@ export default class MenuBot extends Vue {
         return {
             ...this.extraStyle,
             'background-color': this.backgroundColor,
-            height: this.scaleY === 'auto' ? 'auto' : this.scaleY * 40 + 'px',
+            'min-height':
+                this.scaleY === 'auto' ? 'auto' : this.scaleY * 40 + 'px',
         };
     }
 
@@ -168,6 +171,7 @@ export default class MenuBot extends Vue {
             this._updateText(calc, item.bot);
             this._updateCursor(calc, item.bot);
             this._updateAlwaysShowSubmit(calc, item.bot);
+            this._updateInputMultiline(calc, item.bot);
 
             this._updateSim(simulation);
         } else {
@@ -414,6 +418,44 @@ export default class MenuBot extends Vue {
         }
     }
 
+    handleInputEnter(event: KeyboardEvent) {
+        const conditionalSubmit = () => {
+            if (hasValue(this.text)) {
+                this.submitInput(false);
+            }
+        };
+
+        const parsed = Bowser.parse(navigator.userAgent);
+        const isMobile = parsed.platform.type === 'mobile';
+        const isMac = /mac/gi.test(parsed.os.name);
+        if (this.inputMultiline === false) {
+            // always prevent newlines
+            event.preventDefault();
+
+            if (!isMobile) {
+                // On desktop, hitting Enter should submit the input.
+                conditionalSubmit();
+            }
+            return;
+        } else if (this.inputMultiline === 'default') {
+            // allow newlines using shift+enter or on mobile
+            if (!event.shiftKey && !isMobile) {
+                event.preventDefault();
+                conditionalSubmit();
+                return;
+            }
+        }
+
+        if (
+            !event.shiftKey &&
+            ((event.ctrlKey && !isMac) || (isMac && event.metaKey))
+        ) {
+            // If ctrl/cmd is pressed, submit the input.
+            event.preventDefault();
+            conditionalSubmit();
+        }
+    }
+
     async submitInput(dropFocus: boolean) {
         if (dropFocus) {
             const input = <Vue>this.$refs.textInput;
@@ -547,6 +589,16 @@ export default class MenuBot extends Vue {
             'menuItemShowSubmitWhenEmpty',
             false
         );
+    }
+
+    private _updateInputMultiline(calc: BotCalculationContext, bot: Bot) {
+        this.inputMultiline =
+            calculateBooleanTagValue(
+                calc,
+                bot,
+                'auxFormInputMultiline',
+                null
+            ) ?? 'default';
     }
 
     private async _ignoreTextUpdates(action: (text: string) => Promise<void>) {